requirements.txt
ultralytics>=8.0.0
opencv-python>=4.7.0
numpy>=1.24.0
torch>=2.0.0
streamlit>=1.20.0
pillow>=9.0.0
tqdm>=4.64.0
